refactor: move find command building all into the find operation
One of the changes made in merging the core and native drivers was
to merge the wire protocol methods for executing find operations.
A vestigial piece of this was that the find command was being
built in the wire protocol layer, rather than the operation layer.
This patch teases these two apart, so that the `query` wire
protocol method only creates and executes an `OP_QUERY` message,
and the modern and legacy find commands are built in the find
operation's definition itself.

NODE-2900